// src/generated/models/Hr_assignmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Hr_assignmentshr_status = {
  123140000: 'Active',
  123140001: 'Completed'
} as const;
export type Hr_assignmentshr_status = keyof typeof Hr_assignmentshr_status;
export const Hr_assignmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Hr_assignmentsstatecode = keyof typeof Hr_assignmentsstatecode;
export const Hr_assignmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Hr_assignmentsstatuscode = keyof typeof Hr_assignmentsstatuscode;

export interface Hr_assignmentsBase {
  hr_assignmentid: string;
  hr_createdon?: string;
  hr_department?: string;
  "hr_EmailID@odata.bind"?: string;
  hr_executivename?: string;
  hr_expectedreturndate: string;
  hr_modifiedon?: string;
  hr_purpose?: string;
  hr_startdate: string;
  hr_status: Hr_assignmentshr_status;
  "hr_Vehicle@odata.bind": string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Hr_assignmentsstatecode;
  statuscode?: Hr_assignmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Hr_assignments extends Hr_assignmentsBase {
  createdon?: string;
  hr_statusname?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  hr_emailid?: object;
  _hr_emailid_value?: string;
  hr_vehicle?: object;
  _hr_vehicle_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}