// src/generated/models/Hr_usersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Hr_usersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Hr_usersstatecode = keyof typeof Hr_usersstatecode;
export const Hr_usersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Hr_usersstatuscode = keyof typeof Hr_usersstatuscode;

export interface Hr_usersBase {
  hr_department?: string;
  hr_emailid?: string;
  hr_firstname?: string;
  hr_lastname: string;
  hr_userid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Hr_usersstatecode;
  statuscode?: Hr_usersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Hr_users extends Hr_usersBase {
  createdon?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}