// src/generated/models/Hr_vehiclesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Hr_vehicleshr_status = {
  123140000: 'Available',
  123140001: 'Assigned',
  123140002: '123140002'
} as const;
export type Hr_vehicleshr_status = keyof typeof Hr_vehicleshr_status;
export const Hr_vehiclesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Hr_vehiclesstatecode = keyof typeof Hr_vehiclesstatecode;
export const Hr_vehiclesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Hr_vehiclesstatuscode = keyof typeof Hr_vehiclesstatuscode;

export interface Hr_vehiclesBase {
  hr_color?: string;
  hr_createdon?: string;
  hr_model?: string;
  hr_modifiedon?: string;
  hr_status: Hr_vehicleshr_status;
  hr_vehicleid: string;
  hr_year?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Hr_vehiclesstatecode;
  statuscode?: Hr_vehiclesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Hr_vehicles extends Hr_vehiclesBase {
  createdon?: string;
  hr_statusname?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}